naver-ocr-request and response not yet parsing and response to android client

+++ backend/src/main/java/com/eku/eku_ocr_test/form/OcrResponseForm.java
@@ -14,7 +14,6 @@ import java.util.ArrayList;
 public class OcrResponseForm {
     private String version;
     private String requestId;
-    private int timestamp;
+    private long timestamp;
     private ArrayList<ImageResponse> images;
-
 }
